proto: move impl to top level module
